+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+### Added
+- Added AWS transport gem and configuration for check-es-query-* sensu calls to use --transport=AWS
+- Update metrics-es-node-graphite.rb and check-es-node-status.rb for Elasticsearch 2.0
 
 ## [0.5.3] - 2016-04-02
 ### Added
@@ -47,6 +50,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 - Update elasticsearch gem to 1.0.14
 
 ### Added
+- Add check-es-cluster-health that checks health status with elasticsearch gem and can use AWS transport for checks.
 - Add check-es-circuit-breakers.rb, to alert when circuit breakers have been tripped
 
 ## [0.2.0] - 2015-10-15